// src/database/db.types.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type GameSessionStatus = "ACTIVE" | "COMPLETED";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Int8 = ColumnType<string, bigint | number | string, bigint | number | string>;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export interface GameRounds {
  answeredAt: Timestamp | null;
  correctCountryCode: string;
  id: Generated<number>;
  isCorrect: Generated<boolean>;
  round_flagids: string;
  roundNumber: number;
  selectedCountryCode: string | null;
  sessionId: number;
}

export interface GameSessions {
  endedAt: Timestamp | null;
  id: Generated<number>;
  score: Generated<number>;
  startedAt: Generated<Timestamp>;
  status: Generated<GameSessionStatus>;
  totalRounds: Generated<number>;
  userId: number;
}

export interface LeaderboardEntries {
  bestScore: Generated<number>;
  bestTimeMs: Int8 | null;
  updatedAt: Generated<Timestamp>;
  userId: number;
}

export interface Users {
  createdAt: Generated<Timestamp>;
  id: Generated<number>;
  passwordHash: string;
  username: string;
}

export interface DB {
  game_rounds: GameRounds;
  game_sessions: GameSessions;
  leaderboard_entries: LeaderboardEntries;
  users: Users;
}